@@ -13,7 +13,7 @@
         }
     },
     "require": {
-        "php": "^5.5 || ^7.0",
+        "php": "^7.0 || ^5.6",
         "zendframework/zend-stdlib": "^3.0"
     },
     "require-dev": {
@@ -22,8 +22,8 @@
         "zendframework/zend-serializer": "^2.6.1",
         "zendframework/zend-servicemanager": "^2.7.5 || ^3.0.3",
         "zendframework/zend-filter": "^2.6",
-        "phpunit/phpunit": "^4.5",
-        "squizlabs/php_codesniffer": "^2.3.1"
+        "phpunit/phpunit": "^6.0.10 || ^5.7.17",
+        "zendframework/zend-coding-standard": "~1.0.0"
     },
     "suggest": {
         "zendframework/zend-eventmanager": "^2.6.2 || ^3.0, to support aggregate hydrator usage",
@@ -49,5 +49,16 @@
         "psr-4": {
             "ZendTest\\Hydrator\\": "test/"
         }
+    },
+    "scripts": {
+        "check": [
+            "@cs-check",
+            "@test"
+        ],
+        "cs-check": "phpcs",
+        "cs-fix": "phpcbf",
+        "test": "phpunit --colors=always",
+        "test-coverage": "phpunit --colors=always --coverage-clover clover.xml",
+        "upload-coverage": "coveralls -v"
     }
 }
