@@ -9,7 +9,7 @@
 
 namespace ZendTest\Hydrator\Aggregate;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use stdClass;
 use Zend\EventManager\EventManager;
@@ -21,7 +21,7 @@ use Zend\Hydrator\HydratorInterface;
 /**
  * Unit tests for {@see AggregateHydrator}
  */
-class AggregateHydratorTest extends PHPUnit_Framework_TestCase
+class AggregateHydratorTest extends TestCase
 {
     /**
      * @var AggregateHydrator
@@ -36,7 +36,7 @@ class AggregateHydratorTest extends PHPUnit_Framework_TestCase
     /**
      * {@inheritDoc}
      */
-    public function setUp()
+    protected function setUp()
     {
         $this->eventManager = $this->prophesize(EventManager::class);
         $this->hydrator     = new AggregateHydrator();
