@@ -9,19 +9,21 @@
 
 namespace ZendTest\Hydrator\Aggregate;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use stdClass;
+use Zend\EventManager\EventManagerInterface;
 use Zend\Hydrator\Aggregate\ExtractEvent;
 use Zend\Hydrator\Aggregate\HydrateEvent;
 use Zend\Hydrator\Aggregate\HydratorListener;
+use Zend\Hydrator\HydratorInterface;
 
 /**
  * Unit tests for {@see HydratorListener}
  */
-class HydratorListenerTest extends PHPUnit_Framework_TestCase
+class HydratorListenerTest extends TestCase
 {
     /**
-     * @var \Zend\Hydrator\HydratorInterface|\PHPUnit_Framework_MockObject_MockObject
+     * @var HydratorInterface|\PHPUnit_Framework_MockObject_MockObject
      */
     protected $hydrator;
 
@@ -35,9 +37,9 @@ class HydratorListenerTest extends PHPUnit_Framework_TestCase
      *
      * @covers \Zend\Hydrator\Aggregate\HydratorListener::__construct
      */
-    public function setUp()
+    protected function setUp()
     {
-        $this->hydrator = $this->getMock('Zend\Hydrator\HydratorInterface');
+        $this->hydrator = $this->createMock('Zend\Hydrator\HydratorInterface');
         $this->listener = new HydratorListener($this->hydrator);
     }
 
@@ -46,7 +48,7 @@ class HydratorListenerTest extends PHPUnit_Framework_TestCase
      */
     public function testAttach()
     {
-        $eventManager = $this->getMock('Zend\EventManager\EventManagerInterface');
+        $eventManager = $this->createMock('Zend\EventManager\EventManagerInterface');
 
         $eventManager
             ->expects($this->exactly(2))
@@ -71,7 +73,7 @@ class HydratorListenerTest extends PHPUnit_Framework_TestCase
         $hydrated = new stdClass();
         $data     = ['foo' => 'bar'];
         $event    = $this
-            ->getMockBuilder('Zend\Hydrator\Aggregate\HydrateEvent')
+            ->getMockBuilder(HydrateEvent::class)
             ->disableOriginalConstructor()
             ->getMock();
 
@@ -97,7 +99,7 @@ class HydratorListenerTest extends PHPUnit_Framework_TestCase
         $object = new stdClass();
         $data   = ['foo' => 'bar'];
         $event  = $this
-            ->getMockBuilder('Zend\Hydrator\Aggregate\ExtractEvent')
+            ->getMockBuilder(ExtractEvent::class)
             ->disableOriginalConstructor()
             ->getMock();
 
