@@ -9,6 +9,7 @@
 
 namespace ZendTest\Hydrator\Filter;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\Filter\FilterComposite;
 use Zend\Hydrator\Exception\InvalidArgumentException;
 use Zend\Hydrator\Filter\GetFilter;
@@ -21,7 +22,7 @@ use Zend\Hydrator\Filter\NumberOfParameterFilter;
  *
  * @covers \Zend\Hydrator\Filter\FilterComposite
  */
-class FilterCompositeTest extends \PHPUnit_Framework_TestCase
+class FilterCompositeTest extends TestCase
 {
     /**
      * @dataProvider getDataProvider
@@ -35,8 +36,8 @@ class FilterCompositeTest extends \PHPUnit_Framework_TestCase
                 $key = 'foo';
             }
 
-            $this->setExpectedException(
-                InvalidArgumentException::class,
+            $this->expectException(InvalidArgumentException::class);
+            $this->expectExceptionMessage(
                 sprintf(
                     'The value of %s should be either a callable or an ' .
                     'instance of Zend\Hydrator\Filter\FilterInterface',
