@@ -9,6 +9,8 @@
 
 namespace ZendTest\Hydrator\Filter;
 
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\Filter\OptionalParametersFilter;
 
 /**
@@ -16,7 +18,7 @@ use Zend\Hydrator\Filter\OptionalParametersFilter;
  *
  * @covers \Zend\Hydrator\Filter\OptionalParametersFilter
  */
-class OptionalParametersFilterTest extends \PHPUnit_Framework_TestCase
+class OptionalParametersFilterTest extends TestCase
 {
     /**
      * @var OptionalParametersFilter
@@ -26,7 +28,7 @@ class OptionalParametersFilterTest extends \PHPUnit_Framework_TestCase
     /**
      * {@inheritDoc}
      */
-    public function setUp()
+    protected function setUp()
     {
         $this->filter = new OptionalParametersFilter();
     }
@@ -62,7 +64,7 @@ class OptionalParametersFilterTest extends \PHPUnit_Framework_TestCase
 
     public function testTriggersExceptionOnUnknownMethod()
     {
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException(InvalidArgumentException::class);
         $this->filter->filter(__CLASS__ . '::' . 'nonExistingMethod');
     }
 
