@@ -7,12 +7,18 @@
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
-namespace ZendTest\Hdyrator;
+namespace ZendTest\Hydrator;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\ObjectProperty;
 
-class HydratorObjectPropertyTest extends \PHPUnit_Framework_TestCase
+class HydratorObjectPropertyTest extends TestCase
 {
+    /**
+     * @var ObjectProperty
+     */
+    private $hydrator;
+
     public function setUp()
     {
         $this->hydrator = new ObjectProperty();
