@@ -9,6 +9,7 @@
 
 namespace ZendTest\Hydrator\NamingStrategy;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\NamingStrategy\CompositeNamingStrategy;
 use Zend\Hydrator\NamingStrategy\NamingStrategyInterface;
 
@@ -17,12 +18,12 @@ use Zend\Hydrator\NamingStrategy\NamingStrategyInterface;
  *
  * @covers \Zend\Hydrator\NamingStrategy\CompositeNamingStrategy
  */
-class CompositeNamingStrategyTest extends \PHPUnit_Framework_TestCase
+class CompositeNamingStrategyTest extends TestCase
 {
     public function testGetSameNameWhenNoNamingStrategyExistsForTheName()
     {
         $compositeNamingStrategy = new CompositeNamingStrategy([
-            'foo' => $this->getMock('Zend\Hydrator\NamingStrategy\NamingStrategyInterface')
+            'foo' => $this->createMock(NamingStrategyInterface::class)
         ]);
 
         $this->assertEquals('bar', $compositeNamingStrategy->hydrate('bar'));
@@ -32,7 +33,7 @@ class CompositeNamingStrategyTest extends \PHPUnit_Framework_TestCase
     public function testUseDefaultNamingStrategy()
     {
         /* @var $defaultNamingStrategy NamingStrategyInterface|\PHPUnit_Framework_MockObject_MockObject*/
-        $defaultNamingStrategy = $this->getMock('Zend\Hydrator\NamingStrategy\NamingStrategyInterface');
+        $defaultNamingStrategy = $this->createMock(NamingStrategyInterface::class);
         $defaultNamingStrategy->expects($this->at(0))
             ->method('hydrate')
             ->with('foo')
@@ -43,7 +44,7 @@ class CompositeNamingStrategyTest extends \PHPUnit_Framework_TestCase
             ->will($this->returnValue('foo'));
 
         $compositeNamingStrategy = new CompositeNamingStrategy(
-            ['bar' => $this->getMock('Zend\Hydrator\NamingStrategy\NamingStrategyInterface')],
+            ['bar' => $this->createMock(NamingStrategyInterface::class)],
             $defaultNamingStrategy
         );
         $this->assertEquals('Foo', $compositeNamingStrategy->hydrate('foo'));
@@ -52,7 +53,7 @@ class CompositeNamingStrategyTest extends \PHPUnit_Framework_TestCase
 
     public function testHydrate()
     {
-        $fooNamingStrategy = $this->getMock('Zend\Hydrator\NamingStrategy\NamingStrategyInterface');
+        $fooNamingStrategy = $this->createMock(NamingStrategyInterface::class);
         $fooNamingStrategy->expects($this->once())
             ->method('hydrate')
             ->with('foo')
@@ -63,7 +64,7 @@ class CompositeNamingStrategyTest extends \PHPUnit_Framework_TestCase
 
     public function testExtract()
     {
-        $fooNamingStrategy = $this->getMock('Zend\Hydrator\NamingStrategy\NamingStrategyInterface');
+        $fooNamingStrategy = $this->createMock(NamingStrategyInterface::class);
         $fooNamingStrategy->expects($this->once())
             ->method('extract')
             ->with('FOO')
