@@ -9,9 +9,11 @@
 
 namespace ZendTest\Hydrator\NamingStrategy;
 
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\NamingStrategy\MapNamingStrategy;
 
-class MapNamingStrategyTest extends \PHPUnit_Framework_TestCase
+class MapNamingStrategyTest extends TestCase
 {
     public function testHydrateMap()
     {
@@ -34,7 +36,7 @@ class MapNamingStrategyTest extends \PHPUnit_Framework_TestCase
 
     public function testSingleMapInvalidValue()
     {
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException(InvalidArgumentException::class);
         new MapNamingStrategy(['foo' => 3.1415]);
     }
 
